Extract `toc/collection_container.rs`
