+++ lib/storage/src/content_manager/toc/mod.rs
@@ -1,3 +1,4 @@
+mod collection_container;
 mod temp_directories;
 
 use std::cmp::max;
@@ -590,47 +591,6 @@ impl TableOfContent {
     ) -> Result<(), StorageError> {
         let operation = ConsensusOperations::remove_replica(collection_name, shard_id, peer_id);
         proposal_sender.send(operation)
-    }
-
-    fn on_transfer_failure_callback(
-        proposal_sender: Option<OperationSender>,
-    ) -> collection::collection::OnTransferFailure {
-        Arc::new(move |transfer, collection_name, reason| {
-            if let Some(proposal_sender) = &proposal_sender {
-                let operation = ConsensusOperations::abort_transfer(
-                    collection_name.clone(),
-                    transfer.clone(),
-                    reason,
-                );
-                if let Err(send_error) = proposal_sender.send(operation) {
-                    log::error!(
-                        "Can't send proposal to abort transfer of shard {} of collection {}. Error: {}",
-                        transfer.shard_id,
-                        collection_name,
-                        send_error
-                    );
-                }
-            }
-        })
-    }
-
-    fn on_transfer_success_callback(
-        proposal_sender: Option<OperationSender>,
-    ) -> collection::collection::OnTransferSuccess {
-        Arc::new(move |transfer, collection_name| {
-            if let Some(proposal_sender) = &proposal_sender {
-                let operation =
-                    ConsensusOperations::finish_transfer(collection_name.clone(), transfer.clone());
-                if let Err(send_error) = proposal_sender.send(operation) {
-                    log::error!(
-                        "Can't send proposal to complete transfer of shard {} of collection {}. Error: {}",
-                        transfer.shard_id,
-                        collection_name,
-                        send_error
-                    );
-                }
-            }
-        })
     }
 
     fn change_peer_state_callback(
@@ -1441,135 +1401,6 @@ impl TableOfContent {
         self.channel_service.id_to_address.read().clone()
     }
 
-    pub fn collections_snapshot_sync(&self) -> consensus_manager::CollectionsSnapshot {
-        self.general_runtime.block_on(self.collections_snapshot())
-    }
-
-    pub async fn collections_snapshot(&self) -> consensus_manager::CollectionsSnapshot {
-        let mut collections: HashMap<CollectionId, collection_state::State> = HashMap::new();
-        for (id, collection) in self.collections.read().await.iter() {
-            collections.insert(id.clone(), collection.state().await);
-        }
-        consensus_manager::CollectionsSnapshot {
-            collections,
-            aliases: self.alias_persistence.read().await.state().clone(),
-        }
-    }
-
-    pub fn apply_collections_snapshot(
-        &self,
-        data: consensus_manager::CollectionsSnapshot,
-    ) -> Result<(), StorageError> {
-        self.general_runtime.block_on(async {
-            let mut collections = self.collections.write().await;
-
-            for (id, state) in &data.collections {
-                let collection_exists = collections.contains_key(id);
-
-                // Create collection if not present locally
-                if !collection_exists {
-                    let collection_path = self.create_collection_path(id).await?;
-                    let snapshots_path = self.create_snapshots_path(id).await?;
-                    let shard_distribution =
-                        CollectionShardDistribution::from_shards_info(state.shards.clone());
-                    let collection = Collection::new(
-                        id.to_string(),
-                        self.this_peer_id,
-                        &collection_path,
-                        &snapshots_path,
-                        &state.config,
-                        self.storage_config
-                            .to_shared_storage_config(self.is_distributed())
-                            .into(),
-                        shard_distribution,
-                        self.channel_service.clone(),
-                        Self::change_peer_state_callback(
-                            self.consensus_proposal_sender.clone(),
-                            id.to_string(),
-                            ReplicaState::Dead,
-                            None,
-                        ),
-                        Self::request_shard_transfer_callback(
-                            self.consensus_proposal_sender.clone(),
-                            id.to_string(),
-                        ),
-                        Some(self.search_runtime.handle().clone()),
-                        Some(self.update_runtime.handle().clone()),
-                    )
-                    .await?;
-                    collections.validate_collection_not_exists(id).await?;
-                    collections.insert(id.to_string(), collection);
-                }
-
-                let collection = match collections.get(id) {
-                    Some(collection) => collection,
-                    None => unreachable!(),
-                };
-
-                // Update collection state
-                if &collection.state().await != state {
-                    if let Some(proposal_sender) = self.consensus_proposal_sender.clone() {
-                        // In some cases on state application it might be needed to abort the transfer
-                        let abort_transfer = |transfer| {
-                            if let Err(error) =
-                                proposal_sender.send(ConsensusOperations::abort_transfer(
-                                    id.clone(),
-                                    transfer,
-                                    "sender was not up to date",
-                                ))
-                            {
-                                log::error!(
-                                    "Can't report transfer progress to consensus: {}",
-                                    error
-                                )
-                            };
-                        };
-                        collection
-                            .apply_state(state.clone(), self.this_peer_id(), abort_transfer)
-                            .await?;
-                    } else {
-                        log::error!("Can't apply state: single node mode");
-                    }
-                }
-
-                // Mark local shards as dead (to initiate shard transfer),
-                // if collection has been created during snapshot application
-                if !collection_exists {
-                    for shard_id in collection.get_local_shards().await {
-                        collection
-                            .set_shard_replica_state(
-                                shard_id,
-                                self.this_peer_id,
-                                ReplicaState::Dead,
-                                None,
-                            )
-                            .await?;
-                    }
-                }
-            }
-
-            // Remove collections that are present locally but are not in the snapshot state
-            for collection_name in collections.keys() {
-                if !data.collections.contains_key(collection_name) {
-                    log::debug!(
-                        "Deleting collection {collection_name} \
-                         because it is not part of the consensus snapshot",
-                    );
-
-                    self.delete_collection(collection_name).await?;
-                }
-            }
-
-            // Apply alias mapping
-            self.alias_persistence
-                .write()
-                .await
-                .apply_state(data.aliases)?;
-
-            Ok(())
-        })
-    }
-
     pub async fn create_snapshot(
         &self,
         collection_name: &str,
@@ -1675,19 +1506,6 @@ impl TableOfContent {
         Ok(())
     }
 
-    pub async fn remove_shards_at_peer(&self, peer_id: PeerId) -> Result<(), StorageError> {
-        let collections = self.collections.read().await;
-        for collection in collections.values() {
-            collection.remove_shards_at_peer(peer_id).await?;
-        }
-        Ok(())
-    }
-
-    pub fn remove_shards_at_peer_sync(&self, peer_id: PeerId) -> Result<(), StorageError> {
-        self.general_runtime
-            .block_on(self.remove_shards_at_peer(peer_id))
-    }
-
     /// Wait until all other known peers reach the given commit
     ///
     /// # Errors
@@ -1783,98 +1601,5 @@ impl TableOfContent {
             })
             .await
             .map_err(Into::into)
-    }
-}
-
-impl CollectionContainer for TableOfContent {
-    fn perform_collection_meta_op(
-        &self,
-        operation: CollectionMetaOperations,
-    ) -> Result<bool, StorageError> {
-        self.perform_collection_meta_op_sync(operation)
-    }
-
-    fn collections_snapshot(&self) -> consensus_manager::CollectionsSnapshot {
-        self.collections_snapshot_sync()
-    }
-
-    fn apply_collections_snapshot(
-        &self,
-        data: consensus_manager::CollectionsSnapshot,
-    ) -> Result<(), StorageError> {
-        self.apply_collections_snapshot(data)
-    }
-
-    fn remove_peer(&self, peer_id: PeerId) -> Result<(), StorageError> {
-        self.general_runtime.block_on(async {
-            // Validation:
-            // 1. Check that we are not removing some unique shards (removed)
-
-            // Validation passed
-
-            self.remove_shards_at_peer(peer_id).await?;
-
-            if self.this_peer_id == peer_id {
-                // We are detaching the current peer, so we need to remove all connections
-                // Remove all peers from the channel service
-
-                let ids_to_drop: Vec<_> = self
-                    .channel_service
-                    .id_to_address
-                    .read()
-                    .keys()
-                    .filter(|id| **id != self.this_peer_id)
-                    .copied()
-                    .collect();
-                for id in ids_to_drop {
-                    self.channel_service.remove_peer(id).await;
-                }
-            } else {
-                self.channel_service.remove_peer(peer_id).await;
-            }
-            Ok(())
-        })
-    }
-
-    fn sync_local_state(&self) -> Result<(), StorageError> {
-        self.general_runtime.block_on(async {
-            let collections = self.collections.read().await;
-            let transfer_failure_callback =
-                Self::on_transfer_failure_callback(self.consensus_proposal_sender.clone());
-            let transfer_success_callback =
-                Self::on_transfer_success_callback(self.consensus_proposal_sender.clone());
-
-            for collection in collections.values() {
-                let finish_shard_initialize = Self::change_peer_state_callback(
-                    self.consensus_proposal_sender.clone(),
-                    collection.name(),
-                    ReplicaState::Active,
-                    Some(ReplicaState::Initializing),
-                );
-                let convert_to_listener_callback = Self::change_peer_state_callback(
-                    self.consensus_proposal_sender.clone(),
-                    collection.name(),
-                    ReplicaState::Listener,
-                    Some(ReplicaState::Active),
-                );
-                let convert_from_listener_to_active_callback = Self::change_peer_state_callback(
-                    self.consensus_proposal_sender.clone(),
-                    collection.name(),
-                    ReplicaState::Active,
-                    Some(ReplicaState::Listener),
-                );
-
-                collection
-                    .sync_local_state(
-                        transfer_failure_callback.clone(),
-                        transfer_success_callback.clone(),
-                        finish_shard_initialize,
-                        convert_to_listener_callback,
-                        convert_from_listener_to_active_callback,
-                    )
-                    .await?;
-            }
-            Ok(())
-        })
     }
 }
